# api/openapi.json
{
  "openapi": "3.0.0",
  "basePath": "/api",
  "info": {
    "title": "Toll connect",
    "version": "1.0.0",
    "description": "API for SoftEng 2024 NTUA course"
  },
  "definitions": {
...
    "Error": {
      "type": "object",
      "properties": {
        "status": { "type": "string" },
        "info": { "type": "string" }
      },
      "required": ["status", "info"],
      "additionalProperties": true
    },
...
    "UnauthorizedResponse": {
      "description": "Unauthorized ",
      "content": {
        "application/json": { "schema": { "$ref": "#/definitions/Error" } }
      }
    },
    "BadRequestResponse": {
      "description": "Bad Request",
      "content": {
        "application/json": { "schema": { "$ref": "#/definitions/Error" } }
      }
    },
    "InternalServerErrorResponse": {
      "description": "Internal Server Error",
      "content": {
        "application/json": { "schema": { "$ref": "#/definitions/Error" } }
      }
    },
...
    "PassAnalysisResponse": {
      "type": "object",
      "properties": {
        "stationOpID": {
          "type": "string",
          "description": "The ID of the station operator"
        },
        "tagOpID": {
          "type": "string",
          "description": "The ID of the tag operator/provider"
        },
        "requestTimestamp": {
          "type": "string",
          "format": "date-time",
          "description": "Request timestamp"
        },
        "periodFrom": {
          "type": "string",
          "format": "date",
          "description": "Period from (YYYY-MM-DD)"
        },
        "periodTo": {
          "type": "string",
          "format": "date",
          "description": "Period to (YYYY-MM-DD)"
        },
        "nPasses": {
          "type": "integer",
          "description": "Number of pass events"
        },
        "passList": {
          "type": "array",
          "items": { "$ref": "#/definitions/PassAnalysisItem" }
        }
      },
      "required": [
        "stationOpID",
        "tagOpID",
        "requestTimestamp",
        "periodFrom",
        "periodTo",
        "nPasses",
        "passList"
      ]
    },
    "PassAnalysisItem": {
      "type": "object",
      "properties": {
        "passIndex": { "type": "integer", "description": "Serial number" },
        "passID": { "type": "string", "description": "Pass event ID" },
        "stationID": { "type": "string", "description": "Station ID" },
        "timestamp": {
          "type": "string",
          "format": "date-time",
          "description": "Pass timestamp"
        },
        "tagID": { "type": "string", "description": "Tag ID" },
        "passCharge": {
          "type": "number",
          "format": "float",
          "description": "Pass cost"
        }
      },
      "required": [
        "passIndex",
        "passID",
        "stationID",
        "timestamp",
        "tagID",
        "passCharge"
      ]
    },
...
  },
  "paths": {
...
    "/api/passAnalysis/{operator_id}/{tag_id}/{date_from}/{date_to}": {
      "get": {
        "tags": ["Operations"],
        "summary": "Pass Analysis between Operators",
        "description": "Returns an object containing a list of pass events made with tags from tagOpID at stations of stationOpID for the given period.",
        "operationId": "getPassAnalysis",
        "parameters": [
          {
            "name": "operator_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "The ID of the station operator"
          },
          {
            "name": "tag_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "The ID of the tag operator/provider"
          },
          {
            "name": "date_from",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "date" },
            "description": "The start date of the period (YYYYMMDD)"
          },
          {
            "name": "date_to",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "date" },
            "description": "The end date of the period (YYYYMMDD)"
          },
          { "$ref": "#definitions/TokenHeader" }
        ],
        "responses": {
          "200": {
            "description": "Successful retrieval of pass information",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/definitions/PassAnalysisResponse" }
              }
            }
          },
          "400": { "$ref": "#/definitions/BadRequestResponse" },
          "401": { "$ref": "#/definitions/UnauthorizedResponse" },
          "500": { "$ref": "#/definitions/InternalServerErrorResponse" }
        }
      }
    },
...
  }
}